# openapi.snapshot.json
{"components":{"schemas":{"DocumentRecord":{"properties":{"author":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Author"},"doc_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doc Type"},"document_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Document Id"},"group_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Group Id"},"ingested_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ingested At"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language"},"library_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Library Id"},"security_level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Security Level"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"}},"title":"DocumentRecord","type":"object"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail","type":"array"}},"title":"HTTPValidationError","type":"object"},"HealthResponse":{"properties":{"status":{"title":"Status","type":"string"},"tool_id":{"title":"Tool Id","type":"string"}},"required":["status","tool_id"],"title":"HealthResponse","type":"object"},"QueryRequest":{"properties":{"author":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Author"},"catalog_file":{"default":"","title":"Catalog File","type":"string"},"doc_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Doc Type"},"document_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Document Id"},"group_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Group Id"},"ingest_date_from":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"ISO date lower bound (YYYY-MM-DD), inclusive","title":"Ingest Date From"},"ingest_date_to":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"ISO date upper bound (YYYY-MM-DD), inclusive","title":"Ingest Date To"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language"},"library_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Library Id"},"limit":{"default":25,"maximum":500.0,"minimum":1.0,"title":"Limit","type":"integer"},"offset":{"default":0,"minimum":0.0,"title":"Offset","type":"integer"},"security_level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Security Level"},"sort_by":{"default":"document_id","title":"Sort By","type":"string"},"sort_dir":{"default":"asc","title":"Sort Dir","type":"string"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"}},"title":"QueryRequest","type":"object"},"QueryResponse":{"properties":{"limit":{"title":"Limit","type":"integer"},"offset":{"title":"Offset","type":"integer"},"results":{"items":{"$ref":"#/components/schemas/DocumentRecord"},"title":"Results","type":"array"},"sort_by":{"title":"Sort By","type":"string"},"sort_dir":{"title":"Sort Dir","type":"string"},"total":{"title":"Total","type":"integer"}},"required":["results","total","limit","offset","sort_by","sort_dir"],"title":"QueryResponse","type":"object"},"ValidationError":{"properties":{"ctx":{"title":"Context","type":"object"},"input":{"title":"Input"},"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location","type":"array"},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}},"required":["loc","msg","type"],"title":"ValidationError","type":"object"}}},"info":{"title":"document_catalog_query","version":"0.1.0"},"openapi":"3.1.0","paths":{"/v1/tools/document_catalog_query/health":{"get":{"operationId":"document_catalog_query_health","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}},"description":"Successful Response"}},"summary":"Document Catalog Query Health","tags":["tools:document_catalog_query"]}},"/v1/tools/document_catalog_query/query":{"post":{"operationId":"document_catalog_query_query","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryResponse"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Document Catalog Query Query","tags":["tools:document_catalog_query"]}}}}